+++ src/components/Tile.tsx
@@ -1,5 +1,10 @@
 import React from 'react';
 import StyledTile from '@styles/components/Tile.styled';
 
-type TileProps = { children: React.ReactNode };
-export const Tile = (props: TileProps): JSX.Element => <StyledTile>{ props.children }</StyledTile>;
+type TileProps = {
+  children: React.ReactNode;
+  width?: string;
+};
+
+export const Tile = (props: TileProps): JSX.Element =>
+  <StyledTile width={props.width}>{ props.children }</StyledTile>;